+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Work In Progress
+- Addition and subtraction of Tranforms supported to help with easing function calculations
+
+## v0.4.0-alpha0.3
+- Update `golem` to `v0.1.1` to fix non-power-of-2 textures
+- `impl std::iter::Sum for geom::Vector`
+
 ## v0.4.0-alpha0.2
 - Fix the "easy-log" feature
 - [BREAKING] `fill_circle` now takes a `&Circle` instead of a separate center and radius